move fix module into the security fix api module
