+++ uaclient/api/u/pro/security/fix/usn/plan/v1.py
@@ -2,9 +2,9 @@ from typing import List
 
 from uaclient.api.api import APIEndpoint
 from uaclient.api.data_types import AdditionalInfo
+from uaclient.api.u.pro.security.fix import FixPlanUSNResult, fix_plan_usn
 from uaclient.config import UAConfig
 from uaclient.data_types import DataObject, Field, StringDataValue, data_list
-from uaclient.fix import FixPlanUSNResult, fix_plan_usn
 from uaclient.security import FixStatus
